refactor: move suggest API request to suggest package
