+++ core/com.b2international.snowowl.core/META-INF/MANIFEST.MF
@@ -80,6 +80,7 @@ Export-Package: com.auth0.jwt,
  com.b2international.snowowl.core.request.expand,
  com.b2international.snowowl.core.request.io,
  com.b2international.snowowl.core.request.resource,
+ com.b2international.snowowl.core.request.suggest,
  com.b2international.snowowl.core.request.version,
  com.b2international.snowowl.core.scripts,
  com.b2international.snowowl.core.setup,
